[Syntax] make syntax module
