--- lexer/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "lexer"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-chumsky = "0.12.0"
-cst = {path = "../cst"}
-errors = {path = "../errors"}
-location = {path = "../location"}
-pp = {path = "../pp"}
-
-num-bigint = "0.4.6"
