+++ syntax/Cargo.toml
@@ -1,15 +1,13 @@
 [package]
-name = "parser"
+name = "syntax"
 version = "0.1.0"
 edition = "2024"
 
 [dependencies]
 cst = {path = "../cst"}
 errors = {path = "../errors"}
-lexer = {path = "../lexer"}
 location = {path = "../location"}
 pp = {path = "../pp"}
 
 chumsky = "0.12.0"
 num-bigint = "0.4.6"
-
