+++ syntax/src/error.rs
@@ -1,8 +1,31 @@
-use crate::ErrorChumsky;
+use crate::lexer::ErrorChumsky as ErrorLexer;
+use crate::parser::ErrorChumsky as ErrorParser;
 use errors::{Error, Message};
 use location::Span;
 
-pub fn error_parsing(err: &ErrorChumsky<'_>) -> Error {
+pub fn lexing<'src>(err: &ErrorLexer<'src>, offset: usize) -> Error {
+    Error::new(201, Message::text("Lexing error"))
+        .with_span({
+            let span = err.span();
+            Span::new(span.start, span.end).with_offset(offset)
+        })
+        .with_text({
+            let msg = if err.expected().len() == 1 {
+                Message::text("Lexer expected ")
+                    .append(Message::quoted(err.expected().next().unwrap().to_string()))
+                    .append_if(err.found().is_some(), || Message::text(", found "))
+            } else {
+                Message::text("Lexer unknow token ")
+            };
+            msg.append_opt(
+                err.found()
+                    .map(|found| Message::quoted(found.to_string().escape_default()).important()),
+            )
+            .with_text(".")
+        })
+}
+
+pub fn parsing(err: &ErrorParser<'_>) -> Error {
     use chumsky::error::RichPattern;
 
     let expected: Vec<_> = err
