+++ syntax/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod lexer;
+pub mod parser;
+pub mod error;
